fix: restore original migrate.ts, rename our v2 migration to migrate-v2.ts
- Restore src/migrate.ts (v0 -> pro migration by original author)
- Rename our schema-v2 migration to src/migrate-v2.ts to avoid conflict
  (migrate-v2.ts: non-destructive v1→v2 migration adding agentId/tags/redactStatus)
